+++ src/NetCore/Web/Api/wwwroot/scripts/api/ui/webgis_queryResults.js
@@ -124,13 +124,13 @@
                     if (layer != null && layer.selectable === false)
                         selectable = false;
                 }
-
-                webgis.ui.showLayerNotVisibleNotification(service, query, $parent);
             }
             if (query && $.inArray(query.name, includedQueries) < 0) {
                 includedQueries.push(query.name);
             }
         }
+
+        webgis.ui.showLayerNotVisibleNotification(service, query, $parent);
 
         map.getSelection('selection').remove();
         map.getSelection('custom').remove();